# swagger-spec.json
{"openapi":"3.0.0","paths":{"/personagem/criar":{"post":{"operationId":"PersonagemController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreatePersonagemDto"}}}},"responses":{"201":{"description":""}},"summary":"Cadastrar novo personagem","tags":["Personagem"]}},"/personagem":{"get":{"operationId":"PersonagemController_findAll","parameters":[],"responses":{"200":{"description":""}},"summary":"Listar todos os personagem","tags":["Personagem"]}},"/personagem/{id}":{"get":{"operationId":"PersonagemController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"summary":"Listar um personagem por ID","tags":["Personagem"]}},"/personagem/listarItensPersonagem/{personagemID}":{"get":{"operationId":"PersonagemController_listarItensPersonagem","parameters":[{"name":"personagemID","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"summary":"Listar Itens do personagem","tags":["Personagem"]}},"/personagem/listarAmuleto/{personagemID}":{"get":{"operationId":"PersonagemController_listarAmuletoPersonagem","parameters":[{"name":"personagemID","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"summary":"Listar Amuleto do personagem","tags":["Personagem"]}},"/personagem/atualizar/{id}":{"put":{"operationId":"PersonagemController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdatePersonagemDto"}}}},"responses":{"200":{"description":""}},"summary":"Atualizar Personagem por ID do personagem","tags":["Personagem"]}},"/personagem/deletar/{id}":{"delete":{"operationId":"PersonagemController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"summary":"Remover Personagem","tags":["Personagem"]}},"/personagem/deletarItem/{personagemID}/{nomeItem}":{"delete":{"operationId":"PersonagemController_deletarItemPersonagem","parameters":[{"name":"personagemID","required":true,"in":"path","schema":{"type":"string"}},{"name":"nomeItem","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"summary":"Deletar Item do Personagem","tags":["Personagem"]}},"/item-magico/criar":{"post":{"operationId":"ItemMagicoController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateItemMagicoDto"}}}},"responses":{"201":{"description":""}},"summary":"Cadastrar novo Item","tags":["ItemMagico"]}},"/item-magico":{"get":{"operationId":"ItemMagicoController_findAll","parameters":[],"responses":{"200":{"description":""}},"summary":"Listar todos os Itens","tags":["ItemMagico"]}},"/item-magico/{id}":{"get":{"operationId":"ItemMagicoController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"summary":"Listar os Item por ID do Item","tags":["ItemMagico"]}},"/item-magico/adicionarItemAoPersonagem/{nomeItem}/{personagemId}":{"get":{"operationId":"ItemMagicoController_adcionarItem","parameters":[{"name":"nomeItem","required":true,"in":"path","schema":{"type":"string"}},{"name":"personagemId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"summary":"Adicionar Item ao Personagem","tags":["ItemMagico"]}}},"info":{"title":"RPG API","description":"Documentação da API para gerenciamento de personagens e itens mágicos","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"schemas":{"CreatePersonagemDto":{"type":"object","properties":{"nome":{"type":"string"},"nomeAventureiro":{"type":"string"},"classe":{"type":"string"},"level":{"type":"number"},"itemMagico":{"type":"array","items":{"type":"string"}},"forca":{"type":"number"},"defesa":{"type":"number"}},"required":["nome","nomeAventureiro","classe","level","itemMagico","forca","defesa"]},"UpdatePersonagemDto":{"type":"object","properties":{"nome":{"type":"string"},"nomeAventureiro":{"type":"string"},"classe":{"type":"string"},"level":{"type":"number"},"itemMagico":{"type":"array","items":{"type":"string"}},"forca":{"type":"number"},"defesa":{"type":"number"}}},"CreateItemMagicoDto":{"type":"object","properties":{"nome":{"type":"string"},"tipoItem":{"type":"string"},"forca":{"type":"number"},"defesa":{"type":"number"}},"required":["nome","tipoItem","forca","defesa"]}}}}